requirements.txt
Flask>=2.0.0
Pillow>=9.0.0
landingai>=0.1.0
Werkzeug>=2.0.0
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
flask-mail>=0.9.1
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
